we_debug: disable K&R tparm() prototype that conflicts with <term.h>
Comment out the K&R-style 'char *tparm();' declaration inside #ifdef
NOTPARM in we_debug.c. On modern glibc this conflicts with the proper
prototype from <term.h> when NOTPARM is ever defined. NOTPARM is not
set in the Debian build, but the conflict was preventing compilation
on systems where it was.

Origin: Debian package xwpe, debian/patches/tparm_proto